Updated expandable admin page to have individual scoring
